tweak file uploader

+++ website/components/inputs/ImageField.vue
@@ -95,6 +95,8 @@ export default {
         })
     },
     async removeImage() {
+      if(!this.fileDoc) return
+      
       this.uploadPending = true
       fileProvider
         .removeFile(this.fileDoc._id)
